redesign web settings with sidebar layout, add account picker: - replace separate /settings route with sidebar tabs - replace drag-and-drop account list with dropdown account picker - move notification email management to a dedicated tab - desktop: remove account switcher from sidebar footer - desktop: redesign app settings with sidebar layout
